% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cws_groups.R
\name{clean_lvls}
\alias{clean_lvls}
\title{Clean up categories and groups from crosstabs}
\usage{
clean_lvls(x)
}
\arguments{
\item{x}{A vector. If not a factor already, will be coerced to one.}
}
\value{
A factor
}
\description{
This is a bunch of string cleaning to unify the categories (Gender, Age, etc) and groups (Male, Ages 65+, etc) across all available crosstabs. This does the same operation on both categories and groups because there is some overlap; therefore, all warnings are suppressed because it would otherwise be very noisy.
}
