% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_response.R
\name{combine_response}
\alias{combine_response}
\title{Combine multiple response categories into one indicator-style one}
\usage{
combine_response(
  data,
  categories,
  response = response,
  value = value,
  filter_responses = TRUE,
  drop_nonanswers = TRUE,
  nons = c("Don't know", "Refused"),
  ...
)
}
\arguments{
\item{data}{A data frame with groups maintained.}

\item{categories}{A named list of vectors. Each list item should be a character vector of response categories to lump together, and each vector's name should be the new name for the aggregated response. This follows the \code{...} argument of \code{\link[forcats:fct_collapse]{forcats::fct_collapse()}}, as that is where it is passed directly.}

\item{response}{Bare column name of where responses are found, including
those considered to be non-answers. Default: response}

\item{value}{Bare column name of values, Default: value}

\item{filter_responses}{Logical: if \code{TRUE} (the default), only responses matching the names of the \code{categories} argument will be kept.}

\item{drop_nonanswers}{Logical: if \code{TRUE} (the default), will drop nonanswers with \code{\link[=sub_nonanswers]{sub_nonanswers()}} after aggregating.}

\item{nons}{Character vector of responses to be removed.
Default: c("Don't know", "Refused")}

\item{...}{
  Arguments passed on to \code{\link[=sub_nonanswers]{sub_nonanswers}}
  \describe{
    \item{\code{factor_response}}{Logical: if \code{TRUE} (default), returns response variable
as a factor. This is likely a more useful way to handle response
categories once non-answers have been removed.}
    \item{\code{rescale}}{Logical: if \code{TRUE}, values will be scaled based on their total. If \code{FALSE} (the default), values are scaled based on an assumption that all responses add to 1. In some cases, crosstabs with heavy rounding might not add up to 1 when they should, so rescaling helps handle that.}
  }}
}
\value{
A data frame with groups maintained. It will have fewer rows, depending on arguments passed for filtering and dropping nonanswers.
}
\description{
This is a function to speed up a common task of combining multiple responses (e.g. "Excellent", "Good") into a single one (e.g. "Excellent / Good"). It's useful for converting questions in their original format into one where we can use a single number to show the percentage of people who rated something excellent or good. Two additional steps are optional but done by default: non-answers are dropped with \code{\link[=sub_nonanswers]{sub_nonanswers()}}, and the data is filtered for \emph{only} the named aggregated responses. Both of these steps can be disabled independently.
}
\details{
A few gotchas are possible with this function:
\itemize{
\item Unless you're using a data frame with only one location, category, and group, you'll likely want to use this on a grouped data frame. This function \emph{maintains} all groups of the data frame passed in, but it is outside the purview of this function to \emph{create} any groups. You should do that in advance yourself (see examples).
\item Removing non-answers and filtering happen independently but are informed by each other. If you decide to filter the data but not drop non-answers, the values of \code{nons} will be included as response categories to keep. This is to avoid a situation where more data is dropped than you might intend.
}
}
\examples{
  xt <- system.file("extdata/test_xtab2021.xlsx", package = "dcws") |>
      read_xtabs(process = TRUE, year = 2021)
# how responsive is local govt?
  local_govt <- dplyr::filter(xt, code == "Q4A", category == "Age")
  unique(local_govt$response)
# combine excellent & good, drop non-answers, filter
# note that this question has different wording for "Don't know" response
  local_govt |>
     dplyr::group_by(category, group) |>
     combine_response(
         list(excellent_good = c("Excellent", "Good")),
         nons = c("Don't know enough about it in order to say", "Refused")
     )
# combine but no filtering or dropping
local_govt |>
     dplyr::group_by(category, group) |>
     combine_response(
         list(excellent_good = c("Excellent", "Good")),
         nons = c("Don't know enough about it in order to say", "Refused"),
         filter_responses = FALSE,
         drop_nonanswers = FALSE
     )

# multiple combined categories--useful for Likert questions
area_change <- dplyr::filter(xt, code == "Q2", category == "Gender")
unique(area_change$response)
# using default filtering, this drops middle category ("About the same")
area_change |>
  dplyr::group_by(category, group) |>
  combine_response(
     list(
         `Getting better` = c("Much better", "Somewhat better"),
         `Getting worse` = c("Much worse", "Somewhat worse")
     ))
# instead turn off filtering but keep non-answer dropping
# unfortunately now responses are out of order...
area_change |>
  dplyr::group_by(category, group) |>
  combine_response(
     list(
         `Getting better` = c("Much better", "Somewhat better"),
         `Getting worse` = c("Much worse", "Somewhat worse")
     ), filter_responses = FALSE)
# ...unless response is already a factor, in which case levels stick
area_change |>
  dplyr::group_by(category, group) |>
  dplyr::mutate(response = forcats::as_factor(response)) |>
  combine_response(
     list(
         `Getting better` = c("Much better", "Somewhat better"),
         `Getting worse` = c("Much worse", "Somewhat worse")
     ), filter_responses = FALSE)

}
\seealso{
\code{\link[=sub_nonanswers]{sub_nonanswers()}} \code{\link[forcats:fct_collapse]{forcats::fct_collapse()}}

Other data manipulation functions: 
\code{\link{collapse_n_wt}()},
\code{\link{sub_nonanswers}()}
}
\concept{data manipulation functions}
