% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cws_codebook}
\alias{cws_codebook}
\title{DCWS codebook}
\format{
A data frame with 905 rows and 4 variables:
\describe{
\item{year}{Numeric, year of survey}
\item{code}{Character, question code}
\item{question}{Character, full text of survey question}
\item{responses}{List of character vectors, of all possible responses for each question}
}
}
\source{
Compiled DCWS crosstabs
}
\usage{
cws_codebook
}
\description{
This is a data frame of all the codes and questions, and corresponding possible responses, for each year of the survey. Previously this was part of the \code{cws_full_data} data frame, but is now split out on its own to reduce redundancy and save space. Note that surveys are identified only by year, not span; pooled data use codes following their endyears.
}
\examples{
cws_codebook |>
    dplyr::filter(grepl("adequate shelter", question))

cws_codebook |>
    dplyr::filter(grepl("adequate shelter", question), year == 2024) |>
    tidyr::unnest(responses)

}
\keyword{datasets}
