% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cws_demo}
\alias{cws_demo}
\title{DCWS demo data}
\format{
A data frame with 32 rows and 9 variables:
\describe{
\item{year}{Numeric, endyear of survey (e.g. 2024)}
\item{span}{Character, span of years of the survey (e.g. "2015_2024")}
\item{name}{Text of location}
\item{code}{Question code, e.g. "Q2", "Q4E", "RENTEVICT"}
\item{category}{Factor: participant group categories, e.g. "Gender", "Age"}
\item{group}{Factor: participant group, e.g. "Male", "Ages 65+"}
\item{response}{Factor: text of responses, depending on question}
\item{value}{Share of participants giving each response}
\item{weight}{Weights per group for aggregating}
}
}
\source{
Sample of DCWS crosstabs
}
\usage{
cws_demo
}
\description{
This is a small sample of 2018 DataHaven Community Wellbeing Survey data for Greater New Haven with weights attached, saved here for use in examples and testing. It was created with the \code{fetch_cws} function.
}
\examples{
cws_demo
}
\seealso{
\code{\link[=fetch_cws]{fetch_cws()}}
}
\keyword{datasets}
