% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cws_full_wts}
\alias{cws_full_wts}
\title{DCWS weights}
\format{
A data frame with 386 rows and 4 variables:
\describe{
\item{year}{Numeric, year of survey}
\item{span}{Character, span of years of the survey (e.g. "2015_2024")}
\item{name}{Text of location}
\item{weights}{A list of nested data frames, each of which has 2 columns for group and weight.}
}
}
\source{
Compiled DCWS crosstabs
}
\usage{
cws_full_wts
}
\description{
This is a nested data frame containing each survey's weights, used for combining groups to calculate average values. These can be joined to DCWS data with \code{fetch_cws}, or manually. Note that in some larger areas for 2018 (maybe also 2015), groups might not all line up between data and weights--check for NAs in your weights column if need be.
}
\examples{
cws_full_wts

cws_full_wts |>
    dplyr::filter(name == "Greater New Haven") |>
    tidyr::unnest(weights)

}
\seealso{
\code{\link[=fetch_cws]{fetch_cws()}}, \code{\link[=fetch_wts]{fetch_wts()}}
}
\keyword{datasets}
