% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cws_max_moe}
\alias{cws_max_moe}
\title{DCWS maximum margins of error}
\format{
A data frame with 386 rows and 4 variables:
\describe{
\item{year}{Numeric, year of survey}
\item{span}{Character, span of years of the survey (e.g. "2015_2024")}
\item{name}{Text of location}
\item{moe}{Numeric, value of maximum MOE}
}
}
\source{
Compiled DCWS crosstabs
}
\usage{
cws_max_moe
}
\description{
This is a data frame listing the maximum margin of error for estimates for each location and year. The values are extracted from the headers of the crosstab spreadsheets, but many are missing, especially for 2015. It's unlikely anyone will need this often beyond the community profiles on our website, so they're not filtered for the same subset of locations as the main data and weights datasets.
}
\examples{
cws_max_moe

}
\keyword{datasets}
