% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcws-package.R
\docType{package}
\name{dcws-package}
\alias{dcws}
\alias{dcws-package}
\title{dcws: Data from the DataHaven Community Wellbeing Survey}
\description{
This is a collection of data from all past years of the DCWS, plus a bit of metadata and a few functions for extracting subsets of data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ct-data-haven.github.io/dcws}
}

}
\author{
\strong{Maintainer}: Camille Seaberry \email{camille@ctdatahaven.org}

}
\keyword{internal}
