% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cws.R
\name{fetch_wts}
\alias{fetch_wts}
\title{Fetch and subset weights for DCWS data}
\usage{
fetch_wts(..., .year = NULL, .name = NULL, .unnest = FALSE)
}
\arguments{
\item{...}{Any number of conditions to filter by, which will be passed to \code{dplyr::filter}. These don't override the named options, so if you filter by \code{year > 2020} but then set \code{.year = 2015} you're not going to get any data.}

\item{.year}{A vector of one or more year(s) to subset by. If this is a character that contains a separator (\code{"_"}, \code{"-"}, or a space character), it will be assumed to be a span of years, such as for multi-year pooled crosstabs (e.g. \code{"2015_2024"}). Otherwise it's assumed this is a single year of the survey. If \code{NULL}, no filtering is done by year.}

\item{.name}{A vector of one or more strings giving the name(s) to subset by. If \code{NULL}, no filtering is done by name.}

\item{.unnest}{Boolean: should data be returned nested into a column called \code{data}? Defaults to \code{FALSE}.}
}
\value{
A data frame, with either 3 columns (if \code{.unnest = FALSE}) or 4 columns (if \code{.unnest = TRUE}).
\itemize{
\item For \code{.unnest = FALSE}: columns are \code{year}, \code{name}, and a list of nested data frames \code{weights}
\item For \code{.unnest = TRUE}: \code{year}, \code{name}, \code{group}, and \code{weight}
}
}
\description{
This function returns the nested \code{cws_full_wts} data, meant as a counterpart to \code{fetch_cws}.
}
\examples{
# no filtering
fetch_wts()

# weights are generally useful in combination with actual data
# but unless you unnest in advance, this is messy
fetch_cws(code == "Q4E",
    .year = 2021,
    .name = c("Greater New Haven", "New Haven"), .unnest = TRUE
) |>
    dplyr::left_join(fetch_wts(.unnest = TRUE), by = c("year", "span", "name", "group"))

# instead, use fetch_cws with .add_wts = TRUE to get these weights joined for you
fetch_cws(code == "Q4E",
    .year = 2021,
    .name = c("Greater New Haven", "New Haven"),
    .add_wts = TRUE
)
}
\seealso{
\code{\link[=fetch_cws]{fetch_cws()}} \link{cws_full_wts}

Other data accessing functions: 
\code{\link{fetch_cws}()},
\code{\link{read_xtabs}()},
\code{\link{xtab2df}()}
}
\concept{data accessing functions}
