% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_lvls.R
\name{clean_lvls}
\alias{clean_lvls}
\title{Clean up categories and groups from crosstabs}
\usage{
clean_lvls(x)
}
\arguments{
\item{x}{A vector. If not a factor already, will be coerced to one.}
}
\value{
A factor
}
\description{
This is a bunch of string cleaning to standardize the categories (Gender, Age, etc) and groups (Male, Ages 65+, etc) across all available crosstabs. This does the same operation on both categories and groups because there is some overlap. The lists of regex and other replacements aren't exported, but they aren't hidden either: access them at \code{dcws:::to_replace}, \code{dcws:::to_remove}, \code{dcws:::to_recode}, or \code{dcws:::to_collapse} if you need them.
}
\examples{
# vector of strings as read in from crosstabs
categories <- c("Connecticut", "NH Inner Ring", "Gender", "Age",
                "Race/Ethnicity", "Education", "Income", "Children in HH")
levels(clean_lvls(categories))

groups <- c("M", "F", "18-34", "35 to 49", "65 and older",
            "Black/Afr Amer", "African American/Black", "High School",
            "Less than $15,000", "$15,000 to $30,000", "No")
levels(clean_lvls(groups))
}
