% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cws_defs}
\alias{cws_defs}
\title{DCWS indicator definitions}
\format{
A data frame with 33 rows and 3 variables:
\describe{
\item{\code{indicator}}{Text of abbreviated indicator name, e.g. "safe biking"}
\item{\code{question}}{Text of question as given on the survey with punctuation and capital letters removed, e.g. "there are places to bicycle in or near my neighborhood that are safe from traffic such as on the street or on special lanes separate paths or trails"}
\item{\code{collapsed_responses}}{Comma-separated text of responses that are collapsed into the indicator, e.g. "Strongly agree, Somewhat agree".}
}
}
\source{
Handwritten by Camille for the upcoming glossary project
}
\usage{
cws_defs
}
\description{
This data frame is a reference of how indicators are defined, such as Likert questions that get collapsed into a single number (e.g. strongly agree & somewhat agree --> percent agree). It also has more complicated indicators, such as smoking rate and underemployment, and responses used to calculate each. Responses should be consistent across years.
}
\examples{
cws_defs

}
\keyword{datasets}
