% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cws_full_wts}
\alias{cws_full_wts}
\title{DCWS weights}
\format{
A data frame with 97 rows and 3 variables:
\describe{
\item{\code{year}}{Numeric, year of survey}
\item{\code{name}}{Text of location}
\item{\code{weights}}{A list of nested data frames, each of which has 2 columns for group and weight.}
}
}
\source{
Compiled DCWS crosstabs
}
\usage{
cws_full_wts
}
\description{
This is a nested data frame containing each survey's weights, used for combining groups to calculate average values. These can be joined to DCWS data with \code{fetch_cws}, or manually. Note that in some larger areas for 2018 (maybe also 2015), groups might not all line up between data and weights--check for NAs in your weights column if need be.
}
\seealso{
\code{\link[=fetch_cws]{fetch_cws()}}, \code{\link[=fetch_wts]{fetch_wts()}}
}
\keyword{datasets}
