% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cws_group_meta}
\alias{cws_group_meta}
\title{DCWS group metadata}
\format{
A data frame with 82 rows and 3 variables:
\describe{
\item{\code{year}}{Numeric, year of survey}
\item{\code{name}}{Text of location}
\item{\code{groups}}{A list of nested data frames, each of which has 2 columns for category and group.}
}
}
\source{
Compiled DCWS crosstabs
}
\usage{
cws_group_meta
}
\description{
This is a reference dataset listing what categories and groups are available for each survey by year and location. Not all questions are available for all groups.
}
\keyword{datasets}
