% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cws_group_meta}
\alias{cws_group_meta}
\title{DCWS group metadata}
\format{
A data frame with 112 rows and 3 variables:
\describe{
\item{\code{year}}{Numeric, year of survey}
\item{\code{name}}{Text of location}
\item{\code{groups}}{A list of nested data frames, each of which has 2 columns for category and group.}
}
}
\source{
Compiled DCWS crosstabs
}
\usage{
cws_group_meta
}
\description{
This is a reference dataset listing what categories and groups are available for each survey by year and location. Not all questions are available for all groups, and not all groups are available every year or for every location.
}
\examples{
# larger areas have more groups available each year (see number of rows per nested tibble)
cws_group_meta

cws_group_meta |>
  dplyr::filter(name == "Greater New Haven") |>
  tidyr::unnest(groups)

# this is useful if you want to know what locations have data for a certain
# set of conditions, e.g. 2021 values by income
cws_group_meta |>
  tidyr::unnest(groups) |>
  dplyr::filter(year == 2021, category == "Income")

}
\keyword{datasets}
